move rules to own directory
