+++ twitterstream.go
@@ -1,10 +1,13 @@
 // Package twitterstream provides an easy way to stream tweets using Twitter's v2 Streaming API.
 package twitterstream
 
-import "github.com/fallenstedt/twitter-stream/httpclient"
+import (
+	"github.com/fallenstedt/twitter-stream/httpclient"
+	"github.com/fallenstedt/twitter-stream/rules"
+)
 
 type TwitterApi struct {
-	Rules  IRules
+	Rules  rules.IRules
 	Stream IStream
 }
 
@@ -20,7 +23,7 @@ func NewTokenGenerator() ITokenGenerator {
 // It is used to interact with Twitter's v2 filtered streaming API
 func NewTwitterStream(token string) *TwitterApi {
 	client := httpclient.NewHttpClient(token)
-	rules := newRules(client)
+	rules := rules.NewRules(client)
 	stream := newStream(client, newStreamResponseBodyReader())
 	return &TwitterApi{Rules: rules, Stream: stream}
 }
